New test code
These files are code for testing the Arduino in the switch and relay system. switch_state is for testing 1 switch at a time, printing switch on or switch off in the serial monitor. Switch_Reading is for the solar car specifically, checking all the switches and printing which of them are on or that none of them are on in the serial monitor. relay_single_test is just a quick 2 seconds on 1 second off relay code with no sensors controlling it.