Working Lights
Added comments for a lot of the smaller files of code
the Full_lights_attempt is one block of code to get the lights to work, is functional, but messy and a little bugged.
Full_lights_functionalization is a cleaned up version of the code and has multiple files to divide the program into cleaner parts, easier to read and almost completely free of bugs.